Remove redundant file and code
